(PC-17812) favorites: fix imports after moving files
