+++ src/features/favorites/components/FavoritesResults.tsx
@@ -1,5 +1,3 @@
-import { Favorite } from 'features/favorites/atoms/Favorite'
-import { NumberOfResults } from 'features/favorites/atoms/NumberOfResults'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { FlatList } from 'react-native'
 import styled from 'styled-components/native'
@@ -7,7 +5,9 @@ import styled from 'styled-components/native'
 import { FavoriteOfferResponse, FavoriteResponse } from 'api/gen'
 import { BookingOfferModal } from 'features/bookOffer/pages/BookingOfferModal'
 import { Sort } from 'features/favorites/components/Buttons/Sort'
+import { Favorite } from 'features/favorites/components/Favorite'
 import { NoFavoritesResult } from 'features/favorites/components/NoFavoritesResult'
+import { NumberOfResults } from 'features/favorites/components/NumberOfResults'
 import { FavoriteSortBy } from 'features/favorites/pages/FavoritesSorts'
 import { useFavoritesState } from 'features/favorites/pages/FavoritesWrapper'
 import { useFavorites } from 'features/favorites/pages/useFavorites'
